feat(backend): modified controller error handling

+++ backend/controllers/product.controller.js
@@ -62,6 +62,12 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Invalid product ID",
+    });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -70,9 +76,9 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.error("Error in deleting product:", error.message);
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "Product not found",
+      message: "Server error",
     });
   }
 };
